+++ lib/configuration.ts
@@ -1,0 +1,6 @@
+export interface LdapServerMockConfiguration {
+  readonly certPath?: string;
+  readonly certKeyPath?: string;
+  readonly port?: number;
+  readonly searchBase?: string;
+}