+++ lib/user.ts
@@ -1,0 +1,4 @@
+export interface LdapUser {
+  readonly dn: string;
+  readonly attributes: { [key: string]: boolean | string | number | string[] | number[] | boolean[] };
+}